Add files via upload
This paper involves the Railway Track Tracer technology for Creature Detection is a technology that 
detects fractures in railway tracks. This method will assist in preventing numerous train accidents. 
This device routinely checks railway rails with a camera and DC motors to detect fractures and take 
appropriate action to avoid accidents. The Internet of Things is the most studied field, and its 
applications are endless. Internet of Things (IOT) has been implemented to provide up-to-date 
updates on railway management. In this mode, an IR sensor is used to check the availability of the 
platform. This system is used to update the platform availability to the upcoming train to avoid or to 
prevent accidents by the usage of Zigbee, which acts as wireless communication from station to train 
and from train to station and detects fire and automatic engine detachment. To update the platform 
availability. There has been an upsurge in railway accidents caused mainly by track quandaries. That 
might be a misalignment, an obstacle, a crack in one of the track's sides, or some other fault with the 
track. So, keeping track of all these difficulties is a time-consuming chore for a person. The scope of 
the paper encompasses detecting the cracks on rail tracks using network signals - checking the 
platform availability, Zigbee for the communication, fire sensor for detecting the fire ultrasonic 
sensors, and Nodemcu can be used. The Railway Track Tracer technology for Creature Detection is 
a technology that detects fractures in railway tracks. This method will assist in preventing numerous 
train accidents.